--- src/Workspaces/Core/Portable/Workspace/ProjectSystem/FileWatchedAnalyzerReferenceFactory.cs
@@ -1,13 +1,0 @@
-﻿// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-// See the LICENSE file in the project root for more information.
-
-using Microsoft.CodeAnalysis.Diagnostics;
-using Microsoft.CodeAnalysis.Workspaces.ProjectSystem;
-
-namespace Microsoft.CodeAnalysis.ProjectSystem;
-
-internal sealed class FileWatchedAnalyzerReferenceFactory(IFileChangeWatcher fileChangeWatcher)
-    : AbstractFileWatchedReferenceFactory<AnalyzerReference>(fileChangeWatcher)
-{
-}
